refactor(chaosengine): move guidedcli → crud/chaos/guided, delete empty internal/
Renames aegislab/src/internal/chaosengine/guidedcli → aegislab/src/crud/chaos/guided.
Updates package declarations (guidedcli → guided), rewrites the single external
importer (crud/chaos/handler_guided.go: old cechaos alias + import path → guided),
and updates a stale justfile comment that referenced the old handler path.

internal/chaosengine/ still contains resourcetypes/ (no external importers; out of
PR-C scope). internal/ is not empty so rmdir was not attempted.
