+++ aegislab/src/crud/chaos/guided/config.go
@@ -1,4 +1,4 @@
-package guidedcli
+package guided
 
 import (
 	"fmt"
