+++ aegislab/src/crud/chaos/guided/enumerate.go
@@ -1,4 +1,4 @@
-package guidedcli
+package guided
 
 import (
 	"context"
