+++ aegislab/src/crud/chaos/guided/k8s.go
@@ -1,4 +1,4 @@
-package guidedcli
+package guided
 
 import (
 	"context"
