+++ aegislab/src/crud/chaos/guided/next.go
@@ -1,4 +1,4 @@
-package guidedcli
+package guided
 
 import (
 	"fmt"
