+++ aegislab/src/crud/chaos/guided/systems.go
@@ -1,4 +1,4 @@
-package guidedcli
+package guided
 
 import (
 	"context"
